generate_module_files.py: Minor improvements
- Use print(x, file=f) to have newlines at the end of the written files
- Create initial metadata.json file if it doesn't exist
- Rewrite metadata.json only if needed (with newline)
- Improve error handling when repository of the module is not found

+++ scripts/generate_module_files.py
@@ -27,27 +27,38 @@ def generate_source_json(new_version_dir: Path, tarball: str, repo_name: str, mo
     source_dict["strip_prefix"] = f"{repo_name}-{module_version}"
     source_dict["url"] = tarball
     with open(new_version_dir / "source.json", "w") as f:
-        f.write(json.dumps(source_dict, indent=4))
+        print(json.dumps(source_dict, indent=4), file=f)
 
-def change_metadata(module_path: Path, module_name: str, module_version: str): 
+def change_metadata(module_path: Path, module_name: str, module_version: str, repo_name: str):
     metadata_path = module_path / "metadata.json"
 
-    with metadata_path.open("r") as f:
-        metadata = json.load(f)
-    if module_version not in metadata["versions"]:
-        metadata["versions"].append(module_version)
-    with metadata_path.open("w") as f:
-        json.dump(metadata, f, indent=4)
+    if not metadata_path.exists():
+        metadata = {}
+        metadata["homepage"] = f"https://github.com/eclipse-score/{repo_name}"
+        # userid: curl -s https://api.github.com/users/<your-user-name> | jq '.id'
+        metadata["maintainers"] = [{"name": "<tbd>", "github": "<tbd>", "github_user_id": 0}]
+        metadata["repository"] = [f"github:eclipse-score/{repo_name}"]
+        metadata["versions"] = [module_version]
+        metadata["yanked_versions"] = {}
+        module_path.mkdir(parents=True, exist_ok=True)
+        with metadata_path.open("w") as f:
+            print(json.dumps(metadata, indent=4), file=f)
+    else:
+        with metadata_path.open("r") as f:
+            metadata = json.load(f)
+            if module_version not in metadata["versions"]:
+                with metadata_path.open("w") as f:
+                    metadata["versions"].append(module_version)
+                    print(json.dumps(metadata, indent=4), file=f)
 
 def generate_needed_files(module_name: str, module_version: str, bazel_module_file_content: str, tarball: str, repo_name: str):
     module_path = Path("modules") / module_name
-    change_metadata(module_path, module_name, module_version) 
+    change_metadata(module_path, module_name, module_version, repo_name)
     new_version_dir = module_path / module_version
     new_version_dir.mkdir(exist_ok=True)
     with open(new_version_dir / "MODULE.bazel", "w") as f:
        f.write(bazel_module_file_content)
     generate_source_json(new_version_dir, tarball,repo_name, module_version)
-
 
 
 if __name__ == "__main__": 
@@ -62,6 +73,8 @@ if __name__ == "__main__":
     args = parser.parse_args()
     bazel_file_url = args.module_file_url.replace("https://github.com","https://raw.githubusercontent.com").replace("blob", "refs/tags")
     r = requests.get(bazel_file_url)
-    assert r.ok
+    if not r.ok:
+        print(f"Error fetching {bazel_file_url}: {r.text}")
+        exit(1)
     module_text = r.text
     generate_needed_files(args.module_name, args.module_version, module_text, args.tarball, args.repo_name)
